+++ src/Admin/AdminModel.php
@@ -9,6 +9,7 @@ use Seablast\Seablast\SeablastConstant;
 use Seablast\Seablast\SeablastModelInterface;
 use Seablast\Seablast\Superglobals;
 use stdClass;
+use Webmozart\Assert\Assert;
 
 /**
  * Retrieve items from database
@@ -59,7 +60,7 @@ class AdminModel implements SeablastModelInterface
                 $insertable = false;
             } else {
                 $knowledge = (array) $this->tableContent->knowledge();
-                $table =  $knowledge['table'];
+                $table = $knowledge['table'];
                 $columns = $knowledge['columns'];
                 $editable = $knowledge['editable'];
                 $conditionDetails = $knowledge['conditionDetails'];
@@ -71,6 +72,10 @@ class AdminModel implements SeablastModelInterface
                     $this->configuration->getString(SeablastConstant::APP_SELECTED_TABLE),
                     $this->adminHelper->getAllowedTables(SeablastConstant::ADMIN_TABLE_INSERT_ROW)
                 );
+                Assert::isArray($knowledge['booleanLike']);
+                Assert::allString($knowledge['booleanLike']);
+                // TableViewModel produces it, admin.latte consumes it
+                $this->configuration->setArrayString(SeablastConstant::ADMIN_BOOLEAN_FIELDS, $knowledge['booleanLike']);
             }
 
             return (object) [